Reorganize proof
