+++ radix-engine/src/model/substates/mod.rs
@@ -2,6 +2,7 @@ mod component;
 mod kv_store_entry;
 mod non_fungible;
 mod package;
+mod proof;
 mod resource_manager;
 mod substate;
 mod system;
@@ -12,6 +13,7 @@ pub use component::*;
 pub use kv_store_entry::*;
 pub use non_fungible::*;
 pub use package::*;
+pub use proof::*;
 pub use resource_manager::*;
 pub use substate::*;
 pub use system::*;
